+++ CHANGELOG.md
@@ -5,6 +5,10 @@ lemma.
 
 ## Iris master
 
+**Changes in `bi`:**
+
+* Add lemmas `big_sepM2_bupd` and `big_sepM2_fupd`. (by Arnaud Daby-Seesaram)
+
 **Changes in `base_logic`:**
 
 * Change `cinv` construction to add lemma `cinv_acc_1` to access